requirements.txt
albumentations==1.3.1
appdirs==1.4.4
brotlipy==0.7.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==2.0.4
click==8.1.6
colorama==0.4.6
cryptography==41.0.2
docker-pycreds==0.4.0
filelock==3.12.2
fsspec==2023.6.0
gitdb==4.0.10
GitPython==3.1.32
huggingface-hub==0.16.4
idna==3.4
imageio==2.31.1
joblib==1.3.1
lazy_loader==0.3
mkl-fft==1.3.6
mkl-random==1.2.2
mkl-service==2.4.0
networkx==3.1
numpy==1.25.0
opencv-python==4.8.0.74
opencv-python-headless==4.8.0.74
packaging==23.1
pandas==2.0.3
pathtools==0.1.2
Pillow==9.4.0
pip==23.1.2
protobuf==4.23.4
psutil==5.9.5
pycparser==2.21
pyOpenSSL==23.2.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0.1
qudida==0.0.4
regex==2023.6.3
requests==2.29.0
safetensors==0.3.1
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.11.1
sentry-sdk==1.28.1
setproctitle==1.3.2
setuptools==67.8.0
six==1.16.0
smmap==5.0.0
threadpoolctl==3.2.0
tifffile==2023.7.18
tokenizers==0.13.3
torch==1.12.1
torchaudio==0.12.1
torchvision==0.13.1
tqdm==4.65.0
transformers==4.31.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==1.26.16
wandb==0.15.6
wheel==0.38.4
win-inet-pton==1.1.0
